+++ index.js
@@ -3,9 +3,11 @@ const env = process.env.NODE_ENV || 'development';       // if not set, to switc
 const mongoose = require("mongoose");
 const config = require('./config/config')[env];        // add the configurations - 'development' or 'production'
 // const configEnv = config[env]                            // settings of the values form config file - the property [env] which is set by me 
-const app = require('express')();                        // use express and start it as app
 const indexRouter = require("./routes");                 // point to the router routes, no need to write index, as it will be default to search for
-const createRouter = require("./routes/create");
+const authRouter = require("./routes/auth");
+const cubeRouter = require("./routes/cube");
+const accessoryRouter = require("./routes/accessory");
+const app = require('express')();                        // use express and start it as app
 
 mongoose.connect(config.databaseUrl, {
     useNewUrlParser: true,          
@@ -21,7 +23,16 @@ mongoose.connect(config.databaseUrl, {
 require('./config/express')(app);                        // require settings for express, providing argument app
 // require('./routes/index')();                    
 
-app.use("/create", createRouter);
+app.use("/", authRouter);
+app.use("/", accessoryRouter);
+app.use("/", cubeRouter);
 app.use("/", indexRouter);                       //  It would be good to separate and group the routes
+
+
+app.get("*", (req, res) => {                 // It's last (after all app.use's)
+    res.render("404", {
+        title: "Error | Cube worshop"
+    })
+})
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
